Allow user to add/remove QS with one click
Change-Id: If86407f7e949085fe885578a1cd7c12e21a90ab6
Signed-off-by: Jason Edson <jaysonedson@gmail.com>
Signed-off-by: xyyx <xyyx@mail.ru>

SystemUI: Update string for add tiles with one click

In 9.0 it shows a string in the qs panel telling you how to add/remove tiles and after 'SystemUI: Allow user to add/remove QS with one click' it still says "Hold and drag to add tiles". Update string to the correct way.

Change-Id: I2912e9254f9ed14436119addc913fe9907dc7f4e
Signed-off-by: Joey Huab <joey@evolution-x.org>